+++ examples/java-mixin/src/main/java/dev/hmclce/example/javamixin/JavaMixinPlugin.java
@@ -1,4 +1,4 @@
-package dev.hmclnex.example.javamixin;
+package dev.hmclce.example.javamixin;
 
 import org.jackhuang.hmcl.plugin.Plugin;
 import org.jackhuang.hmcl.plugin.PluginContext;