+++ examples/kotlin-helloworld/src/main/kotlin/dev/hmclce/example/kotlinhelloworld/KotlinHelloWorldPlugin.kt
@@ -1,4 +1,4 @@
-package dev.hmclnex.example.kotlinhelloworld
+package dev.hmclce.example.kotlinhelloworld
 
 import com.jfoenix.controls.JFXButton
 import javafx.beans.property.ReadOnlyObjectProperty