+++ examples/offline-unlocker/src/main/java/dev/hmclce/plugin/offlineunlocker/MixinAccountListPage.java
@@ -1,4 +1,4 @@
-package dev.hmclnex.plugin.offlineunlocker;
+package dev.hmclce.plugin.offlineunlocker;
 
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;